fix: fail for g3k_manifest_lookup
Because all the scripts use `set -e` and `g3k_manifest_lookup` use jq/yq — it will return non-zero exit code for situations when the field doesn't exist (in this case — `.global.document_url`).

To mitigate this, the subshell for `documentUrl` need to treat the non-zero code separately (which is kind of hacky), but it will work in this specific case.

Link for extra information on this: https://mywiki.wooledge.org/BashFAQ/105#line-187