added "lifecycle_count" var to terraform plan function.
added a tf plan function "internalBucket" for Argo (needed a separate function because of the environment variable used for arg $1)

added a tf apply "internalBucket" for Argo.

made both apply functions check for a tf state file if the bucket exists. If the state does not exist, a new workspace will be created and the bucket will be imported into state. A terraform plan and apply will also be run. If the state does exist, then a terraform plan and apply will be run to make sure we are up to date.